% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsAgilent.R
\name{peakListInfo.AgilentExport}
\alias{peakListInfo.AgilentExport}
\title{peakListInfo.AgilentExport}
\usage{
peakListInfo.AgilentExport(string, defaultCollapse = ";")
}
\arguments{
\item{string}{character vector to be translated}

\item{defaultCollapse}{only used in case of unknown traces. For these, the
resulting elements are joined together into a single character vector which
can be split by this character}
}
\value{
a data.frame
}
\description{
translates the information string which comes with export of
integration info files (peaklists)
}
\examples{
info <- "D:/MassHunter/Data/Ben/Data0001.d, +ESI TIC Scan Frag=125.0V Data0001.d "
peakListInfo.AgilentExport(info)
info <- "D:/MassHunter/Data/Ben/Data0001.d, +ESI EIC(591.2807) Scan Frag=125.0V Data0001.d "
peakListInfo.AgilentExport(info)

}
