% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsAgilent.R
\name{readPeaklist.AgilentExport.memory}
\alias{readPeaklist.AgilentExport.memory}
\title{readPeaklist.AgilentExport.memory}
\usage{
readPeaklist.AgilentExport.memory(textLines, sep = ",", startsString)
}
\arguments{
\item{textLines}{character vector of the data in Agilent peaklist export
format: first line = description and the rest of the lines is the integration
data. Rownumbers are ignored.}

\item{sep}{defines the separator for the peaklist data. Default is ','}

\item{startsString}{character vector that defines which lines are the start
of a peaklist lines in textLines}
}
\value{
function that generates a list of two objects: first element (info)
is data.frame (with info) and second (data) is a list of data.frame's
}
\description{
function factory that returns a function that takes the data
from a character vector which is in the format of an Agilent peaklist
export (single or multiple peaklist(s)) and returns list of lists with
two elements: data (data.frame) and info (list)
}
\note{
a peaklist = the result from the integration of traces/chromatograms by
the Agilent Masshunter software
}
\examples{
demoFile <- fs::path_package("extdata", "Data0001_MS.csv", package = "readAgilent")
result <- readLines(demoFile)
result <- readPeaklist.AgilentExport.memory(result, startsString = "Data0001")()
result$info
result$data[[1]]
result$data[[3]]
}
