% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsAgilent.R
\name{readPeaklist.AgilentExport}
\alias{readPeaklist.AgilentExport}
\title{readPeaklist.AgilentExport}
\usage{
readPeaklist.AgilentExport(
  filename,
  sep = ",",
  startsString =
    paste0(stringr::str_remove(tools::file_path_sans_ext(basename(filename)), pattern =
    "_.*"), ".d")
)
}
\arguments{
\item{filename}{name of the peaklist file}

\item{sep}{defines the separator for the peaklist data. Default is ','}

\item{startsString}{character vector that defines which lines are the start
of a peaklist lines in the file}
}
\value{
a list of dataElements
}
\description{
function factory that returns a function that takes the data
from a file which is in the format of an Agilent peaklist
export (single or multiple peaklist(s)) and returns list of dataElements
}
\note{
a peaklist = the result from the integration of traces/chromatograms by
the Agilent Masshunter software
}
\examples{
demoFile <- fs::path_package("extdata", "Data0001_MS.csv", package = "readAgilent")
result <- readPeaklist.AgilentExport(demoFile, startsString = "Data0001")()
length(result)
result[[1]]
result[[1]]$info
result[[1]]$data
result[[3]]$info
result[[3]]$data

}
